Adjust connect method to new name.
The resolution order for the host configuration is as follows:

        1. signacdb_host
        2. compmatdb_host
        3. database_host

This means, that by default it is not required to specify a host
for signac.db as long as a database_host is generally defined.
'compmatdb_host' is still reckognized for backwards compatibility.

The database backend is now displayed when calling _str__ or
__repr__ from the database instance.