Make package more robust against PySide import errors.
A bad PySide install may lead to exceptions other than
ImportErrors on import. In this case any attempt to
import the package will fail, even when the GUI is
not even used.

Any error occuring during the attempted import of the PySide
package will now be logged to the debug stream, but otherwise
ignored. Another attempt on import the PySide package will be
made when the user actually starts the GUI.