Expose the reset_statepoint() method of Job publicly.
Since the sp interface already gives the user the ability to change a
job instance's state point variable, it makes sense to expose this
method to simply reset the state point variable pubiclyy to the user
as well.

The name 'reset' instead of 'update' is in accordance with current
naming schemes and is aimed at  making the user aware that this function
will reset and not simply 'update' the state point variable.