Implemnet a view update algorithm.
Creating a linked view will now analyze an existing view  directory and
determine which links are obsolete, which need to be updated and which
need to be created. This reduces the number of required file system link
and unlink operation and enables the user to reuse an existing view tree
structure.

The update algorithm is based on creating a tree data structure
representation of the existing link structure; subsequently all branches
are identified and unlinked. As a final step new links are created.