Fix upsert bug in Collection.replace_one() function.
When using Collection.replace_one() in combination with a simple filter,
that means a filter which consists only of the primary key, the upsert
argument would be ignored and a document would always be upserted.

This is especially problematic since the upsert argument by default is
False.

The issue is now caught by a unit test and fixed.