Implement Job.clear() and Job.reset() methods.
Implement unit tests for both methods.

The Job.clear() method removes all job data, but not the job's workspace
directory. This method does nothing if the workspace directory does not
exist, i.e., the job was not previously initialized.

The Job.reset() method, like Job.clear() removes all job data, but not
the job's workspace directory. This method will initialize the job if it
was not previously initialized, after clearing. This is essentially
equivalent to: job.remove(); job.init()