Implement job selection for signac shell command.
This enables users to preset a jobs and a job variable within the
interpreter namespace of the shell opened by the signac shell command.
The `jobs` variable is a generator over the given selection, the `job`
variable contains a reference to the job if the size of the selection is
exactly one, otherwise it is None.