Implement searching with primary ids and primary id-expressions.
This patch enables selecting documents as part of a Collection with
filter expressions that contain primary ids or primary id expressions.
Previously documents could only be selected by primary-id using the
slicing operator.

This patch also enables expressions that contain primary ids, in
combination with operators $eq, $ne, $in, or $nin. Other operators
invalid in combination with a primary-id key.