Merged in feature/h5store (pull request #73)
This feature enables users to store large data in a similar fashion
like JsonDict. The H5Store is exposed as `job.data` with the exact same interface like `job.document`. Data stored in `job.data` is not directly indexed and is designed for the storage of large data (multiple GBs to TBs).

* Add thread-safe H5Store backend to provide dictionary-like interface for HDF5 files via h5py.
* Officially supported types: built-in types, numpy arrays, and pandas DataFrames natively (pandas types via pytables backend).
* Default H5Store added to Job: Job.data.

Approved-by: Carl Simon Adorf <csadorf@umich.edu>
Approved-by: Bradley Dice <bdice@bradleydice.com>
Approved-by: Vyas Ramasubramani <vramasub@umich.edu>