Implement option to get a project within and above a root directory. (#119)
Implement option to get a project within and above a root directory.

This patch implements the option to "get" a project within a root
directory or above it. That means when get_project(root, search=True),
that is identical to switching into root and calling get_project()
without any arguments.

The main motivation is that it is silly that we need to switch the
working directory in order to be able to search for a project. It will
also simplify the implementation of the "get_job()" feature (#118).