Feature/h5store multiprocessing (#123)
* Implement various multiprocessing unit tests for H5Store.

Tests cover opening of multiple writers and readers within the same and
different processes.

* Open H5Store file in read-only mode for __contains and __len__.

This patch enables the late-specification of the file opening mode (for
internal use only) in a few very limited cases, specificaly for the len
check and the __contains__ check.

The caveat is that the file must exist in both cases for those checks
not to fail with an OSError.

* Do not require HDF5-file to exist for __len__ and __contains__.

This patch ensures that a call to __len__ returns 0 for an instance
of HDF5Store, where the underlying HDF5-file does not exist. Similar a
call to __contains__ will return False if the file does not exist.