Documentation for JSONDict constructor.
I added a docstring to the root-namespace exposed JSONDict class and
moved the `parent` constructor argument to the end of the argument
order.

Having the filename be the first argument is much more natural to users
who would want to use this class stand-alone. All internal calls to this
constructor used `filename=...` syntax, so this change does not require
any other changes to the code base.